fix(all): fixed all CI errors

+++ src/core/core.cpp
@@ -6,18 +6,18 @@
 namespace gr::Core{
 
 void init(const Config& cfg){
-    unsigned int depth, stencil, samples;
-    depth   = static_cast<unsigned int>(gr::Math::Clamp(cfg.depthBits,   16, 32));
-    stencil = static_cast<unsigned int>(gr::Math::Clamp(cfg.stencilBits,  0,  8));
-    samples = static_cast<unsigned int>(gr::Math::Clamp(cfg.msaaSamples,  0,  8));
+    float depth, stencil, samples;
+    depth   = gr::Math::Clamp(static_cast<float>(cfg.depthBits),   16.0f, 32.0f);
+    stencil = gr::Math::Clamp(static_cast<float>(cfg.stencilBits),  0.0f,  8.0f);
+    samples = gr::Math::Clamp(static_cast<float>(cfg.msaaSamples),  0.0f,  8.0f);
 
     if (!glfwInit()) return;
     glfwWindowHint(GLFW_CONTEXT_VERSION_MAJOR, 3);
     glfwWindowHint(GLFW_CONTEXT_VERSION_MINOR, 3);
     glfwWindowHint(GLFW_OPENGL_PROFILE, GLFW_OPENGL_CORE_PROFILE);
-    glfwWindowHint(GLFW_DEPTH_BITS, depth);
-    glfwWindowHint(GLFW_STENCIL_BITS, stencil);
-    glfwWindowHint(GLFW_SAMPLES, samples);
+    glfwWindowHint(GLFW_DEPTH_BITS, static_cast<int>(depth));
+    glfwWindowHint(GLFW_STENCIL_BITS, static_cast<int>(stencil));
+    glfwWindowHint(GLFW_SAMPLES, static_cast<int>(samples));
 
     gr::Time::GetElapsedTimeMS();
 }
